+++ src/WorkflowApp/index.ts
@@ -1,2 +1,0 @@
-export * from 'bpmn-server';
-export {configuration} from './configuration.js';