+++ src/scripts/describe.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { argv } from 'node:process';
 import pug from 'pug';
-import { configuration} from '../WorkflowApp/configuration.js';
+import { configuration} from '../sample-app/configuration.js';
 import { BPMNServer,BPMNAPI, Logger ,Definition} from "bpmn-server";
 const logger = new Logger({ toConsole: false});
 const server = new BPMNServer(configuration, logger, { cron: false });